aws - app-elb:modify-attributes: Filter by attribute first in test
Hah, I guess I didn't see that we already had
c7n.resources.appelb.CheckAttributes.  Oh well.  The only thing that's
maybe not great here is that the ValueFilter based CheckAttributes
expects the value to be a bool (a yaml `true`), while the attributes
modifier likes it to be a literal `true` as a string.